requirements.txt
alembic==1.4.2
aniso8601==8.0.0
argon2-cffi==20.1.0  
async-generator==1.10
attrs==20.3.0        
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.2.1
blinker==1.4
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Excel==0.0.7
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-MySQL==1.5.1
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
ipykernel==5.3.4
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyter-offlinenotebook==0.1.0
jupyterlab-pygments==0.1.2
lml==0.1.0
lxml==4.6.1
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.4
openpyxl==3.0.5
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
prometheus-client==0.9.0
prompt-toolkit==3.0.8
psycopg2==2.8.6
pycparser==2.20
pyecharts-jupyter-installer==0.0.3
pyexcel==0.6.6
pyexcel-handsontable==0.0.2
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
pyexcel-xls==0.6.1
pyexcel-xlsx==0.6.0
Pygments==2.7.2
pymongo==3.11.1
PyMySQL==0.10.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dotenv-0.15.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
pywinpty==0.5.7
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.24.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
terminado==0.9.1
testpath==0.4.4
texttable==1.6.3
tornado==6.1
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
WTForms==2.3.3
xlrd==1.2.0
xlwt==1.3.0